requirements.txt
Django==2.0.2
django-configurations==2.0
django-extensions==2.0.0
djangorestframework==3.7.7
ipython==6.2.1
